requirements.txt
opencv-python
numpy
Pillow
gradio
torch
ultralytics
PyYAML